第三次提交，更新了数据库同时实现了在线预览功能。

+++ src/main/java/com/stu/excel/ReadMarkExcel.java
@@ -38,13 +38,14 @@ public class ReadMarkExcel {
             mark.setTeam(row.getCell(j + 8).toString());
             mark.setCredit(Double.parseDouble(row.getCell(j + 9).toString()));
             mark.setGpa(Double.parseDouble(row.getCell(j + 10).toString()));
-            mark.setReview(Integer.parseInt(row.getCell(j + 11).toString()));
+            mark.setReview((int) Double.parseDouble(row.getCell(j + 11).toString()));
             marks.add(mark);
         }
         return marks;
     }
 
+//
 //    public static void main(String[] args) throws IOException {
-//        readXlsx("E:\\test.xlsx");
+//        readXlsx1("E:\\mark.xlsx");
 //    }
 }
